+++ EasyAccess/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Easy Access Change Log
 
+## 1.0.1 (February 16, 2022)
+
+* Updated to FuryCore 1.4.1
+
 ## 1.0.0 (February 15, 2022)
 
 ### Added