+++ FuryCore/CHANGELOG.md
@@ -1,5 +1,12 @@
 ﻿# FuryCore Change Log
 
+## 1.4.1 (February 16, 2022)
+
+### Fixes
+
+* Fixed ToolbarIcons config not working.
+* Fixed icons being pinned to top when toolbar is in a locked position.
+
 ## 1.4.0 (February 15, 2022)
 
 ### Additions